Split Godot utilities in their own files
