+++ edition/voxel_tool_terrain.cpp
@@ -4,7 +4,7 @@
 #include "../storage/voxel_buffer_gd.h"
 #include "../storage/voxel_metadata_variant.h"
 #include "../terrain/fixed_lod/voxel_terrain.h"
-#include "../util/godot/funcs.h"
+#include "../util/godot/ref_counted.h"
 #include "../util/math/conv.h"
 #include "../util/voxel_raycast.h"
