+++ generators/graph/voxel_graph_node_db.h
@@ -2,7 +2,7 @@
 #define VOXEL_GRAPH_NODE_DB_H
 
 #include "../../util/expression_parser.h"
-#include "../../util/godot/funcs.h" // For String hash
+#include "../../util/godot/string.h" // For String hash
 #include "voxel_generator_graph.h"
 #include "voxel_graph_compiler.h"
 #include "voxel_graph_shader_generator.h"
