+++ streams/region/region_file.cpp
@@ -1,6 +1,6 @@
 #include "region_file.h"
 #include "../../streams/voxel_block_serializer.h"
-#include "../../util/godot/funcs.h"
+#include "../../util/godot/string.h"
 #include "../../util/log.h"
 #include "../../util/profiling.h"
 #include "../../util/string_funcs.h"
