+++ streams/vox/vox_data.h
@@ -2,7 +2,7 @@
 #define VOX_DATA_H
 
 #include "../../util/fixed_array.h"
-#include "../../util/godot/funcs.h"
+#include "../../util/godot/string.h"
 #include "../../util/math/color8.h"
 #include "../../util/math/vector3i.h"
 #include "../../util/memory.h"
