+++ terrain/instancing/voxel_instancer.cpp
@@ -1,7 +1,9 @@
 #include "../../edition/voxel_tool.h"
 #include "../../engine/save_block_data_task.h"
 #include "../../util/container_funcs.h"
-#include "../../util/godot/funcs.h"
+#include "../../util/godot/multimesh.h"
+#include "../../util/godot/node.h"
+#include "../../util/godot/ref_counted.h"
 #include "../../util/math/conv.h"
 #include "../../util/profiling.h"
 #include "../../util/string_funcs.h"
