+++ terrain/variable_lod/voxel_mesh_block_vlt.cpp
@@ -1,6 +1,6 @@
 #include "voxel_mesh_block_vlt.h"
 #include "../../constants/voxel_string_names.h"
-#include "../../util/godot/funcs.h"
+#include "../../util/godot/mesh.h"
 #include "../../util/profiling.h"
 #include "../free_mesh_task.h"
