+++ util/godot/multimesh.h
@@ -1,0 +1,13 @@
+#ifndef ZN_GODOT_MULTIMESH_H
+#define ZN_GODOT_MULTIMESH_H
+
+class MultiMesh;
+
+namespace zylann {
+
+// This API can be confusing so I made a wrapper
+int get_visible_instance_count(const MultiMesh &mm);
+
+} // namespace zylann
+
+#endif // ZN_GODOT_MULTIMESH_H
