+++ util/godot/node.cpp
@@ -1,0 +1,28 @@
+#include "node.h"
+#include <scene/main/node.h>
+
+namespace zylann {
+
+template <typename F>
+void for_each_node_depth_first(Node *parent, F f) {
+	ERR_FAIL_COND(parent == nullptr);
+	f(parent);
+	for (int i = 0; i < parent->get_child_count(); ++i) {
+		for_each_node_depth_first(parent->get_child(i), f);
+	}
+}
+
+void set_nodes_owner(Node *root, Node *owner) {
+	for_each_node_depth_first(root, [owner](Node *node) { //
+		node->set_owner(owner);
+	});
+}
+
+void set_nodes_owner_except_root(Node *root, Node *owner) {
+	ERR_FAIL_COND(root == nullptr);
+	for (int i = 0; i < root->get_child_count(); ++i) {
+		set_nodes_owner(root->get_child(i), owner);
+	}
+}
+
+} // namespace zylann
